requirements.txt
python-socketio
python-socketio[client]
requests
pyautogui